reorganize imports and update type definitions for economic data components
