+++ src/modules/expedientes/editar-datos-economicos/components/adjudicacion-form.component.tsx
@@ -1,10 +1,9 @@
-import * as innerClasses from './adjudicacion-form.styles';
-
-import { MontosIvaForm } from './montos-iva-form.component';
 import React from 'react';
-import { TextFieldForm } from '#common/components/index.ts';
 import { Typography } from '@mui/material';
-import { useWithTheme } from '#core/theme/theme.hooks.ts';
+import { TextFieldForm } from '#common/components';
+import { useWithTheme } from '#core/theme';
+import { MontosIvaForm } from './montos-iva-form.component';
+import * as innerClasses from './adjudicacion-form.styles';
 
 export const AdjudicacionForm: React.FC = () => {
   const classes = useWithTheme(innerClasses);
