+++ src/modules/expedientes/editar-datos-economicos/components/presupuesto-base-form.component.tsx
@@ -1,8 +1,8 @@
-import { Anualidad } from '../editar-expediente.type';
-import { ListadoAnualidad } from './listado-anualidad.component';
-import { MontosIvaForm } from './montos-iva-form.component';
 import React from 'react';
 import { Typography } from '@mui/material';
+import { ListadoAnualidad } from './listado-anualidad.component';
+import { MontosIvaForm } from './montos-iva-form.component';
+import { Anualidad } from '../editar-expediente.vm';
 
 export const PresupuestoBaseForm: React.FC<{ rows: Anualidad[] }> = props => {
   const { rows } = props;
