--- src/modules/expedientes/editar-datos-economicos/editar-expediente.type.ts
@@ -1,11 +1,0 @@
-export interface Expediente {
-  id: number;
-}
-
-export interface Anualidad {
-  id: number;
-  ano: number;
-  monto: number;
-  iva: number;
-  simbolo: string;
-}
