+++ src/modules/expedientes/editar-datos-economicos/editar-expediente.vm.ts
@@ -1,3 +1,13 @@
-import { Expediente } from './editar-expediente.type';
+export interface Expediente {
+  id: number;
+}
+
+export interface Anualidad {
+  id: number;
+  ano: number;
+  monto: number;
+  iva: number;
+  simbolo: string;
+}
 
 export const createEmptyExpediente = (): Expediente => ({ id: 0 });
